Add new class add common functions in grant handlers
